+++ src/main/java/com/student/book_advisor/entityRepositories/AuthorJoinBookRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface AuthorJoinBookRepository extends JpaRepository<AuthorJoinBook, Long> {
+public interface AuthorJoinBookRepository extends JpaRepository<AuthorJoinBook, Integer> {
 
     public AuthorJoinBook findByAuthorAndBook(Author author, Libro book);
 }
